Move Parser to piper.parser and Storage to piper.storage
