+++ src/main/java/piper/storage/Storage.java
@@ -1,5 +1,6 @@
-package piper;
+package piper.storage;
 
+import piper.PiperException;
 import piper.task.TaskList;
 import piper.task.Task;
 import piper.task.Todo;
